+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 * No show premium link when it's installed
+* RESOLVE #34 Non-SSL protocol used for image inclusion
 * Use aihr_check_aihrus_framework
 
 ## 2.17.0